Update server.R
The addition of stopApp() will make sure that when user closes the window when running the session locally, the shiny app and the server will be terminated. It is most visible when running from R Console or from Terminal using Rscript, which are still busy even after user closes the browser tab.